#3 [타이머] 타이머 버튼 개발
- 타이머 UI 버튼 개발
