--- app/src/main/java/cookcook/nexters/com/amoogye/views/timer/TimerFragment.kt
@@ -1,29 +1,0 @@
-package cookcook.nexters.com.amoogye.views.timer
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import cookcook.nexters.com.amoogye.R
-
-
-class TimerFragment: Fragment() {
-
-    companion object {
-        // 선택 선언 1 (Fragment를 싱글턴으로 사용 시)
-        private var INSTANCE: TimerFragment? = null
-
-        fun getInstance(): TimerFragment {
-            if (INSTANCE == null) {
-                INSTANCE = TimerFragment()
-            }
-            return INSTANCE!!
-        }
-    }
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        super.onCreateView(inflater, container, savedInstanceState)
-        return inflater.inflate(R.layout.fragment_timer, container, false)
-    }
-}
