+++ app/src/main/res/drawable/timer_button_wrap_rounded_box.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+       android:shape="rectangle">
+    <solid
+            android:width="2dp"
+            android:color="@color/white"/>
+    <corners
+            android:topLeftRadius="10dp"
+            android:topRightRadius="10dp"/>
+</shape>
